pyproject.toml
[tool.poetry]
name = "dwhservice"
version = "0.1.0"
description = "Service which responsible for work with DWH"
authors = ["ALena Meshcheriakova"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
sqlalchemy = "^2.0.32"
logger = "^1.4"
alembic = "^1.13.2"
pytest = "^8.3.2"
pydantic-settings = "^2.4.0"
psycopg2-binary = "^2.9.9"
pytest-dotenv = "^0.5.2"
python-dotenv = "^1.0.1"
pika = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
